+++ Application/ResearchDataManagementPlatform/WindowManagement/ActivateItems.cs
@@ -207,11 +207,10 @@ namespace ResearchDataManagementPlatform.WindowManagement
             HistoryProvider.Add(args.Request.ObjectToEmphasise);
         }
 
-        public override bool SelectEnum(string prompt, Type enumType, out Enum chosen)
+        public override bool SelectEnum(DialogArgs args, Type enumType, out Enum chosen)
         {
-            var selector = new SelectDialog<Enum>(this,Enum.GetValues(enumType).Cast<Enum>().ToArray(), false,false);
-            selector.Text = prompt;
-
+            var selector = new SelectDialog<Enum>(args,this,Enum.GetValues(enumType).Cast<Enum>().ToArray(), false,false);
+            
             if (selector.ShowDialog() == DialogResult.OK)
             {
                 chosen = selector.Selected;
@@ -222,10 +221,9 @@ namespace ResearchDataManagementPlatform.WindowManagement
             return false;
         }
 
-        public override bool SelectType(string prompt, Type[] available,out Type chosen)
+        public override bool SelectType(DialogArgs args, Type[] available,out Type chosen)
         {
-            var dlg =  new SelectDialog<Type>(this,available, false,false);
-            dlg.Text = prompt;
+            var dlg =  new SelectDialog<Type>(args,this,available, false,false);
 
             if (dlg.ShowDialog() == DialogResult.OK)
             {
@@ -476,9 +474,9 @@ namespace ResearchDataManagementPlatform.WindowManagement
         }
 
         /// <inheritdoc/>
-        public override bool YesNo(string text,string caption,out bool chosen)
+        public override bool YesNo(DialogArgs args,out bool chosen)
         {
-            var dr = MessageBox.Show(text, caption, MessageBoxButtons.YesNo);
+            var dr = MessageBox.Show(args.TaskDescription, args.WindowTitle, MessageBoxButtons.YesNo);
 
             if (dr == DialogResult.Yes)
             {
@@ -496,9 +494,9 @@ namespace ResearchDataManagementPlatform.WindowManagement
             return false;
         }
 
-        public override bool TypeText(string header, string prompt, int maxLength, string initialText, out string text, bool requireSaneHeaderText)
+        public override bool TypeText(DialogArgs args, int maxLength, string initialText, out string text, bool requireSaneHeaderText)
         {
-            var textTyper = new TypeTextOrCancelDialog(header, prompt, maxLength, initialText, allowBlankText: false, multiLine: maxLength > 1000)
+            var textTyper = new TypeTextOrCancelDialog(args, maxLength, initialText, allowBlankText: false, multiLine: maxLength > 1000)
             {
                 RequireSaneHeaderText = requireSaneHeaderText
             };
@@ -554,30 +552,22 @@ namespace ResearchDataManagementPlatform.WindowManagement
         }
 
         
-        public override IMapsDirectlyToDatabaseTable SelectOne(string prompt, IMapsDirectlyToDatabaseTable[] availableObjects,
-            string initialSearchText = null, bool allowAutoSelect = false)
+        public override IMapsDirectlyToDatabaseTable SelectOne(DialogArgs args, IMapsDirectlyToDatabaseTable[] availableObjects)
         {
             if (!availableObjects.Any())
             {
-                MessageBox.Show("There are no compatible objects in your RMDP for '"+ prompt +"''");
+                MessageBox.Show($"There are no compatible objects in your RMDP for:{Environment.NewLine}{args}");
                 return null;
             }
 
             //if there is only one object available to select
             if (availableObjects.Length == 1)
-                if(allowAutoSelect || YesNo("You only have one compatible object, use '"+availableObjects[0]+"'","Select '" + availableObjects[0] + "'?"))
+                if(args.AllowAutoSelect)
                 {
                     return availableObjects[0];
                 }
-                else
-                {
-                    return null;
-                }
 
-            var selectDialog = new SelectDialog<IMapsDirectlyToDatabaseTable>(this, availableObjects, false, false);
-            selectDialog.Text = prompt;
-            selectDialog.SetInitialFilter(initialSearchText);
-
+            var selectDialog = new SelectDialog<IMapsDirectlyToDatabaseTable>(args,this, availableObjects, false, false);
 
             if (selectDialog.ShowDialog() == DialogResult.OK)
                 return selectDialog.Selected;
@@ -585,10 +575,9 @@ namespace ResearchDataManagementPlatform.WindowManagement
             return null; //user didn't select one of the IMapsDirectlyToDatabaseTable objects shown in the dialog
         }
 
-        public override bool SelectObject<T>(string prompt, T[] available, out T selected, string initialSearchText = null, bool allowAutoSelect = false)
+        public override bool SelectObject<T>(DialogArgs args, T[] available, out T selected)
         {
-            var pick = new SelectDialog<T>(this,available,false,false);
-            pick.Text = prompt;
+            var pick = new SelectDialog<T>(args,this,available,false,false);
 
             if (pick.ShowDialog() == DialogResult.OK)
             {
@@ -600,10 +589,9 @@ namespace ResearchDataManagementPlatform.WindowManagement
             return false;
         }
 
-        public override bool SelectObjects<T>(string prompt, T[] available, out T[] selected, string initialSearchText = null)
+        public override bool SelectObjects<T>(DialogArgs args, T[] available, out T[] selected)
         {
-            var pick = new SelectDialog<T>(this, available, false, false);
-            pick.Text = prompt;
+            var pick = new SelectDialog<T>(args,this, available, false, false);
             pick.AllowMultiSelect = true;
 
             if (pick.ShowDialog() == DialogResult.OK)
@@ -659,11 +647,10 @@ namespace ResearchDataManagementPlatform.WindowManagement
             }
         }
 
-        protected override bool SelectValueTypeImpl(string prompt, Type paramType, object initialValue, out object chosen)
+        protected override bool SelectValueTypeImpl(DialogArgs args, Type paramType, object initialValue, out object chosen)
         {
             //whatever else it is use string
-            var typeTextDialog = new TypeTextOrCancelDialog("Enter Value", prompt + " (" + paramType.Name + ")",1000,
-                initialValue?.ToString());
+            var typeTextDialog = new TypeTextOrCancelDialog(args,1000,initialValue?.ToString());
 
             if (typeTextDialog.ShowDialog() == DialogResult.OK)
             {
@@ -675,8 +662,8 @@ namespace ResearchDataManagementPlatform.WindowManagement
             return false;
         }
 
-        public override IMapsDirectlyToDatabaseTable[] SelectMany(string prompt, Type arrayElementType,
-            IMapsDirectlyToDatabaseTable[] availableObjects, string initialSearchText)
+        public override IMapsDirectlyToDatabaseTable[] SelectMany(DialogArgs args, Type arrayElementType,
+            IMapsDirectlyToDatabaseTable[] availableObjects)
         {
             if (!availableObjects.Any())
             {
@@ -684,9 +671,7 @@ namespace ResearchDataManagementPlatform.WindowManagement
                 return null;
             }
 
-            var selectDialog = new SelectDialog<IMapsDirectlyToDatabaseTable>(this, availableObjects, false, false);
-            selectDialog.Text = prompt;
-            selectDialog.SetInitialFilter(initialSearchText);
+            var selectDialog = new SelectDialog<IMapsDirectlyToDatabaseTable>(args, this, availableObjects, false, false);
             selectDialog.AllowMultiSelect = true;
             
             if (selectDialog.ShowDialog() == DialogResult.OK)
@@ -779,9 +764,9 @@ namespace ResearchDataManagementPlatform.WindowManagement
             return true;
         }
 
-        public override void SelectAnythingThen(string prompt, Action<IMapsDirectlyToDatabaseTable> callback)
+        public override void SelectAnythingThen(DialogArgs args, Action<IMapsDirectlyToDatabaseTable> callback)
         {
-            NavigateToObjectUI navigate = new NavigateToObjectUI(this) { Text = prompt };
+            NavigateToObjectUI navigate = new NavigateToObjectUI(this) { Text = args.WindowTitle };
             navigate.CompletionAction = callback;
             navigate.Show();
         }