+++ Rdmp.Core.Tests/CommandLine/Interactive/ConsoleInputManagerTests.cs
@@ -8,6 +8,7 @@ using System;
 using System.Collections.Generic;
 using System.Text;
 using NUnit.Framework;
+using Rdmp.Core.CommandExecution;
 using Rdmp.Core.CommandLine.Interactive;
 using ReusableLibraryCode.Checks;
 using Tests.Common;
@@ -22,7 +23,7 @@ namespace Rdmp.Core.Tests.CommandLine.Interactive
             var manager = new ConsoleInputManager(RepositoryLocator, new ThrowImmediatelyCheckNotifier());
             manager.DisallowInput = true;
             
-            Assert.Throws<InputDisallowedException>(()=>manager.GetString("bob", null));
+            Assert.Throws<InputDisallowedException>(()=>manager.GetString(new DialogArgs { WindowTitle = "bob" }, null));
         }
     }
 }