+++ Rdmp.Core/CommandExecution/IBasicActivateItems.cs
@@ -89,14 +89,6 @@ namespace Rdmp.Core.CommandExecution
         /// </summary>
         event EmphasiseItemHandler Emphasise;
 
-
-        /// <summary>
-        /// Show all objects in RDMP (with search).  If a single selection is made then invoke the callback
-        /// </summary>
-        /// <param name="prompt"></param>
-        /// <param name="callback"></param>
-        void SelectAnythingThen(string prompt, Action<IMapsDirectlyToDatabaseTable> callback);
-        
         /// <summary>
         /// Show some SQL and the data that it returns.  This should be non modal
         /// </summary>
@@ -148,6 +140,7 @@ namespace Rdmp.Core.CommandExecution
         /// <returns></returns>
         IPipelineRunner GetPipelineRunner(IPipelineUseCase useCase, IPipeline pipeline);
 
+        #region Select X Modal methods
         /// <summary>
         /// Prompts the user to enter a description for a cohort they are trying to create including whether it is intended to replace an old version of another cohort.
         /// </summary>
@@ -156,6 +149,16 @@ namespace Rdmp.Core.CommandExecution
         /// <param name="cohortInitialDescription">Optional initial description for the cohort which may be changed by the user</param>
         /// <returns></returns>
         CohortCreationRequest GetCohortCreationRequest(ExternalCohortTable externalCohortTable, IProject project, string cohortInitialDescription);
+
+        /// <summary>
+        /// Show all objects in RDMP (with search).  If a single selection is made then invoke the callback
+        /// </summary>
+        /// <param name="prompt"></param>
+        /// <param name="callback"></param>
+        void SelectAnythingThen(string prompt, Action<IMapsDirectlyToDatabaseTable> callback);
+
+        /// <inheritdoc cref="SelectAnythingThen(string, Action{IMapsDirectlyToDatabaseTable})"/>
+        void SelectAnythingThen(DialogArgs args, Action<IMapsDirectlyToDatabaseTable> callback);
 
         /// <summary>
         /// Prompts the user to pick from one of the <paramref name="availableObjects"/> one or more objects.  Returns null or empty if
@@ -168,16 +171,9 @@ namespace Rdmp.Core.CommandExecution
         /// <returns></returns>
         IMapsDirectlyToDatabaseTable[] SelectMany(string prompt, Type arrayElementType,IMapsDirectlyToDatabaseTable[] availableObjects,string initialSearchText = null);
         
-        /// <summary>
-        /// Prompts user or directly creates a new satelite database (e.g. logging / dqe etc) and returns a persistent reference to it
-        /// </summary>
-        /// <param name="catalogueRepository">The main catalogue database</param>
-        /// <param name="defaultToSet">If the created database is to become the new default database of it's type provide this</param>
-        /// <param name="db">The server in which the database should be created or null if the user is expected to pick themselves as part of the method e.g. through a UI</param>
-        /// <param name="patcher">The schema and patches to run to create the database</param>
-        /// <returns></returns>
-        ExternalDatabaseServer CreateNewPlatformDatabase(ICatalogueRepository catalogueRepository, PermissableDefaults defaultToSet, IPatcher patcher, DiscoveredDatabase db);
-         
+        /// <inheritdoc cref="SelectMany(string, Type, IMapsDirectlyToDatabaseTable[], string)"/>
+        IMapsDirectlyToDatabaseTable[] SelectMany(DialogArgs args, Type arrayElementType, IMapsDirectlyToDatabaseTable[] availableObjects);
+
         /// <summary>
         /// Prompts user to pick one of the <paramref name="availableObjects"/>
         /// </summary>
@@ -187,6 +183,13 @@ namespace Rdmp.Core.CommandExecution
         /// <param name="allowAutoSelect"></param>
         /// <returns></returns>
         IMapsDirectlyToDatabaseTable SelectOne(string prompt, IMapsDirectlyToDatabaseTable[] availableObjects, string initialSearchText = null, bool allowAutoSelect = false);
+
+        /// <summary>
+        /// Prompts user to pick one of the <paramref name="availableObjects"/>
+        /// </summary>
+        /// <param name="args"></param>
+        /// <param name="availableObjects">Objects that can be selected</param>
+        IMapsDirectlyToDatabaseTable SelectOne(DialogArgs args, IMapsDirectlyToDatabaseTable[] availableObjects);
 
         /// <summary>
         /// Prompts user to pick one of <paramref name="available"/> objects
@@ -200,6 +203,8 @@ namespace Rdmp.Core.CommandExecution
         /// <returns>True if a selection was made</returns>
         bool SelectObject<T>(string prompt, T[] available, out T selected, string initialSearchText = null, bool allowAutoSelect = false) where T : class;
 
+        /// <inheritdoc cref="SelectObject{T}(string, T[], out T, string, bool)"/>
+        bool SelectObject<T>(DialogArgs args, T[] available, out T selected) where T : class;
 
         /// <summary>
         /// Prompts user to pick several of <paramref name="available"/> objects
@@ -210,6 +215,9 @@ namespace Rdmp.Core.CommandExecution
         /// <param name="selected"></param>
         /// <param name="initialSearchText"></param>
         bool SelectObjects<T>(string prompt, T[] available, out T[] selected, string initialSearchText = null) where T : class;
+
+        /// <inheritdoc cref="SelectObjects{T}(string, T[], out T[], string)"/>
+        bool SelectObjects<T>(DialogArgs args, T[] available, out T[] selected) where T : class;
 
         /// <summary>
         /// Prompts user to select a directory on disk
@@ -244,19 +252,6 @@ namespace Rdmp.Core.CommandExecution
         /// <returns></returns>
         FileInfo SelectFile(string prompt,string patternDescription, string pattern);
         
-        /// <summary>
-        /// Return all Types of the given {T} which should be <see cref="IMapsDirectlyToDatabaseTable"/>
-        /// </summary>
-        /// <typeparam name="T"></typeparam>
-        /// <returns></returns>
-        IEnumerable<T> GetAll<T>();
-
-        /// <summary>
-        ///  Return all Types of the given <paramref name="t"/> which should be <see cref="IMapsDirectlyToDatabaseTable"/>
-        /// </summary>
-        /// <param name="t"></param>
-        /// <returns></returns>
-        IEnumerable<IMapsDirectlyToDatabaseTable> GetAll(Type t);
 
         /// <summary>
         /// User must supply a basic value type e.g. string, double, int
@@ -268,12 +263,10 @@ namespace Rdmp.Core.CommandExecution
         /// <returns></returns>
         bool SelectValueType(string prompt, Type paramType, object initialValue,out object chosen);
 
-        /// <summary>
-        /// Delete the <paramref name="deleteable"/> ideally asking the user for confirmation first (if appropriate)
-        /// </summary>
-        /// <param name="deleteable"></param>
-        /// <returns></returns>
-        bool DeleteWithConfirmation(IDeleteable deleteable);
+
+        /// <inheritdoc cref="SelectValueType(string, Type, object, out object)"/>
+        bool SelectValueType(DialogArgs args, Type paramType, object initialValue, out object chosen);
+
 
         /// <summary>
         /// Offers the user a binary choice and returns true if they consciously select a value.  This method is blocking.
@@ -284,6 +277,9 @@ namespace Rdmp.Core.CommandExecution
         /// <returns>true if user successfully made a choice</returns>
         bool YesNo(string text, string caption, out bool chosen);
 
+        /// <inheritdoc cref="YesNo(string, string, out bool)"/>
+        bool YesNo(DialogArgs args, out bool chosen);
+
         /// <summary>
         /// Offers the user a binary choice returning the choice or false if user cancels (to distinguish between false and cancel
         /// use the overload).  This method is blocking.</summary>
@@ -291,6 +287,10 @@ namespace Rdmp.Core.CommandExecution
         /// <param name="caption"></param>
         /// <returns>Users choice or false if user cancels</returns>
         bool YesNo(string text, string caption);
+
+        /// <inheritdoc cref="YesNo(string, string)"/>
+        bool YesNo(DialogArgs args);
+
 
         /// <summary>
         /// Offers the user a binary choice for whether they want to proceede on a dangerous/complex operation.  Returns true if 
@@ -301,16 +301,8 @@ namespace Rdmp.Core.CommandExecution
         /// <returns></returns>
         bool Confirm(string text, string caption);
 
-        /// <summary>
-        /// Component for auditing errors that should be brought to the users attention subtly (e.g. if a plugin crashes while attempting to create menu items)
-        /// </summary>
-        ICheckNotifier GlobalErrorCheckNotifier { get; }
-        /// <summary>
-        /// Called when <see cref="BasicCommandExecution.Publish"/> is invoked.  Allows you to respond to publish events outside of UI code.  UI code
-        /// should invoke the RefreshBus system in Rdmp.UI
-        /// </summary>
-        /// <param name="databaseEntity"></param>
-        void Publish(IMapsDirectlyToDatabaseTable databaseEntity);
+        /// <inheritdoc cref="Confirm(string, string)"/>
+        bool Confirm(DialogArgs args);
 
         /// <summary>
         /// Display the given message to the user (e.g. in a MessageBox or out into the Console)
@@ -336,6 +328,9 @@ namespace Rdmp.Core.CommandExecution
         /// <param name="requireSaneHeaderText"></param>
         /// <returns></returns>
         bool TypeText(string header, string prompt, int maxLength, string initialText, out string text, bool requireSaneHeaderText);
+
+        /// <inheritdoc cref="TypeText(string, string, int, string, out string, bool)"/>
+        bool TypeText(DialogArgs args, int maxLength, string initialText, out string text, bool requireSaneHeaderText);
 
         /// <summary>
         /// Prompts the user to pick a database
@@ -370,19 +365,6 @@ namespace Rdmp.Core.CommandExecution
         /// <param name="cts"></param>
         void Wait(string title, Task task, CancellationTokenSource cts);
 
-        
-        /// <summary>
-        /// Requests that the activator highlight or otherwise emphasise the supplied item.  Depending on who is subscribed to this event nothing may actually happen
-        /// </summary>
-        void RequestItemEmphasis(object sender, EmphasiseRequest emphasiseRequest);
-
-        /// <summary>
-        /// Returns the root object in the tree hierarchy or the inputted parameter (<paramref name="objectToEmphasise"/>)
-        /// </summary>
-        /// <param name="objectToEmphasise"></param>
-        /// <returns></returns>
-        object GetRootObjectOrSelf(IMapsDirectlyToDatabaseTable objectToEmphasise);
-
         /// <summary>
         /// Requests a selection of one of the values of the <see cref="Enum"/> <paramref name="enumType"/>
         /// </summary>
@@ -391,6 +373,9 @@ namespace Rdmp.Core.CommandExecution
         /// <param name="chosen"></param>
         /// <returns></returns>
         bool SelectEnum(string prompt, Type enumType, out Enum chosen);
+
+        /// <inheritdoc cref="SelectEnum(string, Type, out Enum)"/>
+        bool SelectEnum(DialogArgs args, Type enumType, out Enum chosen);
 
         /// <summary>
         /// Requests user select a <see cref="Type"/> (optional one that is assignable to <paramref name="baseTypeIfAny"/>).
@@ -401,6 +386,9 @@ namespace Rdmp.Core.CommandExecution
         /// <param name="chosen"></param>
         /// <returns></returns>
         bool SelectType(string prompt, Type baseTypeIfAny, out Type chosen);
+
+        /// <inheritdoc cref="SelectValueType(string, Type, object, out object)"/>
+        bool SelectType(DialogArgs args, Type baseTypeIfAny, out Type chosen);
 
 
         /// <summary>
@@ -414,6 +402,9 @@ namespace Rdmp.Core.CommandExecution
         /// <returns></returns>
         bool SelectType(string prompt, Type baseTypeIfAny, bool allowAbstract, bool allowInterfaces, out Type chosen);
 
+        /// <inheritdoc cref="SelectType(string, Type, bool, bool, out Type)"/>
+        bool SelectType(DialogArgs args, Type baseTypeIfAny, bool allowAbstract, bool allowInterfaces, out Type chosen);
+
         /// <summary>
         /// Requests user select one of the <paramref name="available"/> <see cref="Type"/>
         /// </summary>
@@ -422,6 +413,11 @@ namespace Rdmp.Core.CommandExecution
         /// <param name="chosen">The users choice (unless the option is cancelled - see return value)</param>
         /// <returns>True if a choice was made or False if the choice was cancelled</returns>
         bool SelectType(string prompt, Type[] available, out Type chosen);
+
+        /// <inheritdoc cref="SelectType(string, Type[], out Type)"/>
+        bool SelectType(DialogArgs args, Type[] available, out Type chosen);
+
+        #endregion
 
         /// <summary>
         /// Launches an appropriate user interface for <paramref name="o"/> (or does nothing if
@@ -447,5 +443,60 @@ namespace Rdmp.Core.CommandExecution
         /// <param name="cic"></param>
         /// <returns>True if a wizard was shown even if no configuration was then created</returns>
         bool ShowCohortWizard(out CohortIdentificationConfiguration cic);
+
+
+        /// <summary>
+        /// Requests that the activator highlight or otherwise emphasise the supplied item.  Depending on who is subscribed to this event nothing may actually happen
+        /// </summary>
+        void RequestItemEmphasis(object sender, EmphasiseRequest emphasiseRequest);
+
+        /// <summary>
+        /// Returns the root object in the tree hierarchy or the inputted parameter (<paramref name="objectToEmphasise"/>)
+        /// </summary>
+        /// <param name="objectToEmphasise"></param>
+        /// <returns></returns>
+        object GetRootObjectOrSelf(IMapsDirectlyToDatabaseTable objectToEmphasise);
+
+
+        /// <summary>
+        /// Prompts user or directly creates a new satelite database (e.g. logging / dqe etc) and returns a persistent reference to it
+        /// </summary>
+        /// <param name="catalogueRepository">The main catalogue database</param>
+        /// <param name="defaultToSet">If the created database is to become the new default database of it's type provide this</param>
+        /// <param name="db">The server in which the database should be created or null if the user is expected to pick themselves as part of the method e.g. through a UI</param>
+        /// <param name="patcher">The schema and patches to run to create the database</param>
+        /// <returns></returns>
+        ExternalDatabaseServer CreateNewPlatformDatabase(ICatalogueRepository catalogueRepository, PermissableDefaults defaultToSet, IPatcher patcher, DiscoveredDatabase db);
+
+        /// <summary>
+        /// Return all Types of the given {T} which should be <see cref="IMapsDirectlyToDatabaseTable"/>
+        /// </summary>
+        /// <typeparam name="T"></typeparam>
+        /// <returns></returns>
+        IEnumerable<T> GetAll<T>();
+
+        /// <summary>
+        ///  Return all Types of the given <paramref name="t"/> which should be <see cref="IMapsDirectlyToDatabaseTable"/>
+        /// </summary>
+        /// <param name="t"></param>
+        /// <returns></returns>
+        IEnumerable<IMapsDirectlyToDatabaseTable> GetAll(Type t);
+        /// <summary>
+        /// Delete the <paramref name="deleteable"/> ideally asking the user for confirmation first (if appropriate)
+        /// </summary>
+        /// <param name="deleteable"></param>
+        /// <returns></returns>
+        bool DeleteWithConfirmation(IDeleteable deleteable);
+
+        /// <summary>
+        /// Component for auditing errors that should be brought to the users attention subtly (e.g. if a plugin crashes while attempting to create menu items)
+        /// </summary>
+        ICheckNotifier GlobalErrorCheckNotifier { get; }
+        /// <summary>
+        /// Called when <see cref="BasicCommandExecution.Publish"/> is invoked.  Allows you to respond to publish events outside of UI code.  UI code
+        /// should invoke the RefreshBus system in Rdmp.UI
+        /// </summary>
+        /// <param name="databaseEntity"></param>
+        void Publish(IMapsDirectlyToDatabaseTable databaseEntity);
     }
 }