+++ Rdmp.UI/SimpleDialogs/TypeTextOrCancelDialog.Designer.cs
@@ -28,87 +28,101 @@
         /// </summary>
         private void InitializeComponent()
         {
+            this.components = new System.ComponentModel.Container();
             this.textBox1 = new System.Windows.Forms.TextBox();
-            this.label1 = new System.Windows.Forms.Label();
             this.btnOk = new System.Windows.Forms.Button();
             this.btnCancel = new System.Windows.Forms.Button();
-            this.lblNewLineInstructions = new System.Windows.Forms.Label();
+            this.panel1 = new System.Windows.Forms.Panel();
+            this.pTextEditor = new System.Windows.Forms.Panel();
+            this.taskDescriptionLabel1 = new Rdmp.UI.SimpleDialogs.TaskDescriptionLabel();
+            this.toolTip = new System.Windows.Forms.ToolTip(this.components);
+            this.panel1.SuspendLayout();
+            this.pTextEditor.SuspendLayout();
             this.SuspendLayout();
             // 
             // textBox1
             // 
-            this.textBox1.Anchor = ((System.Windows.Forms.AnchorStyles)(((System.Windows.Forms.AnchorStyles.Top | System.Windows.Forms.AnchorStyles.Left) 
-            | System.Windows.Forms.AnchorStyles.Right)));
-            this.textBox1.Location = new System.Drawing.Point(18, 35);
+            this.textBox1.Dock = System.Windows.Forms.DockStyle.Fill;
+            this.textBox1.Location = new System.Drawing.Point(10, 0);
             this.textBox1.Margin = new System.Windows.Forms.Padding(4, 3, 4, 3);
-            this.textBox1.Multiline = true;
             this.textBox1.Name = "textBox1";
-            this.textBox1.Size = new System.Drawing.Size(492, 22);
+            this.textBox1.Size = new System.Drawing.Size(719, 23);
             this.textBox1.TabIndex = 0;
             this.textBox1.TextChanged += new System.EventHandler(this.textBox1_TextChanged);
             this.textBox1.KeyDown += new System.Windows.Forms.KeyEventHandler(this.textBox1_KeyDown);
             // 
-            // label1
-            // 
-            this.label1.Anchor = ((System.Windows.Forms.AnchorStyles)(((System.Windows.Forms.AnchorStyles.Top | System.Windows.Forms.AnchorStyles.Left) 
-            | System.Windows.Forms.AnchorStyles.Right)));
-            this.label1.Location = new System.Drawing.Point(18, 10);
-            this.label1.Margin = new System.Windows.Forms.Padding(4, 0, 4, 0);
-            this.label1.Name = "label1";
-            this.label1.Size = new System.Drawing.Size(492, 21);
-            this.label1.TabIndex = 1;
-            this.label1.Text = "Type some text:";
-            // 
             // btnOk
             // 
             this.btnOk.Anchor = System.Windows.Forms.AnchorStyles.Bottom;
-            this.btnOk.Location = new System.Drawing.Point(170, 69);
+            this.btnOk.Location = new System.Drawing.Point(273, 4);
             this.btnOk.Margin = new System.Windows.Forms.Padding(4, 3, 4, 3);
             this.btnOk.Name = "btnOk";
             this.btnOk.Size = new System.Drawing.Size(88, 27);
             this.btnOk.TabIndex = 2;
             this.btnOk.Text = "Ok";
+            this.toolTip.SetToolTip(this.btnOk, "Press to Save (ENTER)");
             this.btnOk.UseVisualStyleBackColor = true;
             this.btnOk.Click += new System.EventHandler(this.btnOk_Click);
             // 
             // btnCancel
             // 
             this.btnCancel.Anchor = System.Windows.Forms.AnchorStyles.Bottom;
-            this.btnCancel.Location = new System.Drawing.Point(264, 69);
+            this.btnCancel.Location = new System.Drawing.Point(369, 4);
             this.btnCancel.Margin = new System.Windows.Forms.Padding(4, 3, 4, 3);
             this.btnCancel.Name = "btnCancel";
             this.btnCancel.Size = new System.Drawing.Size(88, 27);
             this.btnCancel.TabIndex = 2;
             this.btnCancel.Text = "Cancel";
+            this.toolTip.SetToolTip(this.btnCancel, "Press to Close (ESCAPE)");
             this.btnCancel.UseVisualStyleBackColor = true;
             this.btnCancel.Click += new System.EventHandler(this.btnCancel_Click);
             // 
-            // lblNewLineInstructions
+            // panel1
             // 
-            this.lblNewLineInstructions.Anchor = ((System.Windows.Forms.AnchorStyles)((System.Windows.Forms.AnchorStyles.Bottom | System.Windows.Forms.AnchorStyles.Right)));
-            this.lblNewLineInstructions.AutoSize = true;
-            this.lblNewLineInstructions.Location = new System.Drawing.Point(354, 75);
-            this.lblNewLineInstructions.Name = "lblNewLineInstructions";
-            this.lblNewLineInstructions.Size = new System.Drawing.Size(159, 15);
-            this.lblNewLineInstructions.TabIndex = 3;
-            this.lblNewLineInstructions.Text = "(\'CTRL + Enter\' for New Line)";
-            this.lblNewLineInstructions.Visible = false;
+            this.panel1.Controls.Add(this.btnOk);
+            this.panel1.Controls.Add(this.btnCancel);
+            this.panel1.Dock = System.Windows.Forms.DockStyle.Bottom;
+            this.panel1.Location = new System.Drawing.Point(0, 96);
+            this.panel1.Name = "panel1";
+            this.panel1.Size = new System.Drawing.Size(739, 35);
+            this.panel1.TabIndex = 4;
+            // 
+            // pTextEditor
+            // 
+            this.pTextEditor.Controls.Add(this.textBox1);
+            this.pTextEditor.Dock = System.Windows.Forms.DockStyle.Fill;
+            this.pTextEditor.Location = new System.Drawing.Point(0, 42);
+            this.pTextEditor.Name = "pTextEditor";
+            this.pTextEditor.Padding = new System.Windows.Forms.Padding(10, 0, 10, 0);
+            this.pTextEditor.Size = new System.Drawing.Size(739, 54);
+            this.pTextEditor.TabIndex = 6;
+            // 
+            // taskDescriptionLabel1
+            // 
+            this.taskDescriptionLabel1.AutoSize = true;
+            this.taskDescriptionLabel1.AutoSizeMode = System.Windows.Forms.AutoSizeMode.GrowAndShrink;
+            this.taskDescriptionLabel1.Dock = System.Windows.Forms.DockStyle.Top;
+            this.taskDescriptionLabel1.Location = new System.Drawing.Point(0, 0);
+            this.taskDescriptionLabel1.Name = "taskDescriptionLabel1";
+            this.taskDescriptionLabel1.Size = new System.Drawing.Size(739, 42);
+            this.taskDescriptionLabel1.TabIndex = 7;
             // 
             // TypeTextOrCancelDialog
             // 
             this.AutoScaleDimensions = new System.Drawing.SizeF(7F, 15F);
             this.AutoScaleMode = System.Windows.Forms.AutoScaleMode.Font;
-            this.ClientSize = new System.Drawing.Size(524, 111);
-            this.Controls.Add(this.label1);
-            this.Controls.Add(this.lblNewLineInstructions);
-            this.Controls.Add(this.btnCancel);
-            this.Controls.Add(this.btnOk);
-            this.Controls.Add(this.textBox1);
+            this.ClientSize = new System.Drawing.Size(739, 131);
+            this.Controls.Add(this.pTextEditor);
+            this.Controls.Add(this.panel1);
+            this.Controls.Add(this.taskDescriptionLabel1);
             this.Margin = new System.Windows.Forms.Padding(4, 3, 4, 3);
-            this.MinimumSize = new System.Drawing.Size(540, 150);
+            this.MinimumSize = new System.Drawing.Size(540, 170);
             this.Name = "TypeTextOrCancelDialog";
             this.StartPosition = System.Windows.Forms.FormStartPosition.CenterScreen;
             this.Text = "Input";
+            this.panel1.ResumeLayout(false);
+            this.pTextEditor.ResumeLayout(false);
+            this.pTextEditor.PerformLayout();
             this.ResumeLayout(false);
             this.PerformLayout();
 
@@ -117,9 +131,11 @@
         #endregion
 
         private System.Windows.Forms.TextBox textBox1;
-        private System.Windows.Forms.Label label1;
         private System.Windows.Forms.Button btnOk;
         private System.Windows.Forms.Button btnCancel;
-        private System.Windows.Forms.Label lblNewLineInstructions;
+        private System.Windows.Forms.Panel panel1;
+        private System.Windows.Forms.Panel pTextEditor;
+        private TaskDescriptionLabel taskDescriptionLabel1;
+        private System.Windows.Forms.ToolTip toolTip;
     }
 }