+++ Rdmp.UI/SimpleDialogs/TypeTextOrCancelDialog.cs
@@ -8,7 +8,9 @@ using System;
 using System.Drawing;
 using System.Windows.Forms;
 using FAnsi.Discovery;
+using Rdmp.Core.CommandExecution;
 using Rdmp.UI.ScintillaHelper;
+using ScintillaNET;
 
 namespace Rdmp.UI.SimpleDialogs
 {
@@ -20,47 +22,89 @@ namespace Rdmp.UI.SimpleDialogs
     {
         private readonly bool _allowBlankText;
         private readonly bool _multiline;
-        public string ResultText {get { return textBox1.Text.Trim(); }}
+        private Scintilla _scintilla;
+
+        public string ResultText {get { return (_multiline ? _scintilla.Text : textBox1.Text)?.Trim(); }}
 
         /// <summary>
         /// True to require that text typed be sane for usage as a column name, table name etc e.g. "bob" but not "bob::bbbbb".
         /// </summary>
-        public bool RequireSaneHeaderText{get;set;}
+        public bool RequireSaneHeaderText{get;set; }
 
-        public TypeTextOrCancelDialog(string header, string label, int maxCharacters, string startingTextForInputBox = null, bool allowBlankText = false, bool multiLine = false)
+        //"Column Name","Enter name for column (this should NOT include any qualifiers e.g. database name)", 300);
+
+        public TypeTextOrCancelDialog(string title, string prompt, int maxCharacters, string startingTextForInputBox = null, bool allowBlankText = false, bool multiLine = false)
+            : this(new DialogArgs
+            {
+                WindowTitle = title,
+                EntryLabel = prompt,
+            },maxCharacters,startingTextForInputBox,allowBlankText,multiLine)
+        {
+            
+        }
+
+        public TypeTextOrCancelDialog(DialogArgs args, int maxCharacters, string startingTextForInputBox = null, bool allowBlankText = false, bool multiLine = false)
         {
             _allowBlankText = allowBlankText;
             _multiline = multiLine;
 
             InitializeComponent();
-            
-            if(header.Length > WideMessageBox.MAX_LENGTH_TITLE)
+
+            var header = args.WindowTitle;
+           
+
+            if (header != null && header.Length > WideMessageBox.MAX_LENGTH_TITLE)
                 header = header.Substring(0, WideMessageBox.MAX_LENGTH_TITLE);
 
-            if (label.Length > WideMessageBox.MAX_LENGTH_BODY)
-                label = label.Substring(0, WideMessageBox.MAX_LENGTH_BODY);
+            taskDescriptionLabel1.SetupFor(args);
 
             this.Text = header;
-            this.label1.Text = label;
             this.textBox1.MaxLength = maxCharacters;
-
-            textBox1.Text = startingTextForInputBox;
 
             if (_multiline)
             {
-                lblNewLineInstructions.Visible = true;
+                var editor = new ScintillaTextEditorFactory();
+                _scintilla = editor.Create(null,SyntaxLanguage.None,null,true,false);
+                _scintilla.Dock = DockStyle.Fill;
+                _scintilla.TextChanged += _scintilla_TextChanged;
+                _scintilla.KeyDown += _scintilla_KeyDown;
+                _scintilla.Text = startingTextForInputBox;
+                _scintilla.WrapMode = WrapMode.Word;
+
+                pTextEditor.Controls.Remove(textBox1);
+                pTextEditor.Controls.Add(_scintilla);
+
+                //Move cursor to the end of the textbox
+                this.ActiveControl = _scintilla;
+                _scintilla.SelectionStart = _scintilla.TextLength;
+
                 this.textBox1.Anchor = (AnchorStyles.Top | AnchorStyles.Bottom | AnchorStyles.Left | AnchorStyles.Right);
                 this.textBox1.ScrollBars = ScrollBars.Vertical;
                 this.Height = 290;
                 this.Width = 740;
+
+                //Update the tooltip for the OK button
+                toolTip.SetToolTip(btnOk, "Press to Save (SHIFT + ENTER)");
             }
             else
             {
-                var desiredWidth = TextRenderer.MeasureText(label, label1.Font).Width;
-                Width = Math.Max(540, Math.Min(740, desiredWidth));
+                textBox1.Text = startingTextForInputBox;
+                Width = Math.Max(540, Math.Min(740, taskDescriptionLabel1.PreferredWidth));
+
+                this.ActiveControl = textBox1;
             }
 
             SetEnabledness();          
+        }
+
+        private void _scintilla_KeyDown(object sender, KeyEventArgs e)
+        {
+            FinishedKeyCheck(e);
+        }
+
+        private void _scintilla_TextChanged(object sender, EventArgs e)
+        {
+            SetEnabledness();
         }
 
         private void btnOk_Click(object sender, EventArgs e)
@@ -77,24 +121,7 @@ namespace Rdmp.UI.SimpleDialogs
 
         private void textBox1_KeyDown(object sender, KeyEventArgs e)
         {
-            //If they've pressed enter...
-            if (e.KeyCode == Keys.Enter)
-            {
-                if(!(_multiline && e.Control))
-                {
-                    //Supress the enter key (so a new line isn't created) and press the OK button (if it's enabled) if we're NOT (in a multiline control while holding the control key)
-                    //i.e. in a multiline control hold CTRL+Enter to create a new line, else we treat enter as normal and "OK" the window
-                    e.Handled = true;
-                    e.SuppressKeyPress = true;
-
-                    if (btnOk.Enabled)
-                        btnOk_Click(null, null);
-                }
-            }
-
-            //Escape should work for all controls
-            if (e.KeyCode == Keys.Escape)
-                btnCancel_Click(null, null);
+            FinishedKeyCheck(e);
         }
 
         private void textBox1_TextChanged(object sender, EventArgs e)
@@ -118,7 +145,28 @@ namespace Rdmp.UI.SimpleDialogs
                 }
             }
 
-            btnOk.Enabled = (!string.IsNullOrWhiteSpace(textBox1.Text)) || _allowBlankText;
+            btnOk.Enabled = (!string.IsNullOrWhiteSpace(ResultText)) || _allowBlankText;
+        }
+
+        private void FinishedKeyCheck(KeyEventArgs e)
+        {
+            //If they've pressed enter...
+            if (e.KeyCode == Keys.Enter)
+            {
+                //If the OK button is enabled AND... (we're not multiline OR we are multiline but they're holding shift)
+                if(btnOk.Enabled && (!_multiline || (_multiline && e.Shift)))
+                {
+                    //Supress the enter key (so a new line isn't created) and press the OK button
+                    e.Handled = true;
+                    e.SuppressKeyPress = true;
+                    btnOk_Click(null, null);
+                }
+            }
+
+            //Escape should work for all controls
+            if (e.KeyCode == Keys.Escape)
+             btnCancel_Click(null, null);
+
         }
     }
 }